Split parseArgs and printHelps functions
